+++ rotation_3d.c
@@ -6,7 +6,7 @@
 /*   By: lguiller <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/26 15:39:47 by lguiller          #+#    #+#             */
-/*   Updated: 2018/10/23 13:54:22 by lguiller         ###   ########.fr       */
+/*   Updated: 2018/11/20 16:10:29 by lguiller         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -54,7 +54,7 @@ t_mat3	ft_vecrotz(t_mat3 v, double a)
 	return (w);
 }
 
-t_mat3	ft_vecrot(t_mat3 a, float xrot, float yrot, float zrot)
+t_mat3	ft_vecrot(t_mat3 a, double xrot, double yrot, double zrot)
 {
 	t_mat3	v;
